Support bootloaders larger than 0x10000 bytes.
Any boot code destined for address 0x10000 or higher is now written to
dynamically allocated 64kb pages. The memory map allows this to continue
as high as address 0x3fffffff. That whole region is also used for RAM
and IO, so it may need more adjustments and dynamic page allocations
later.